LayoutTests/discovery/get-network-services-upnp-events-expected.txt

Tests navigator.getNetworkServices() works in happy case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() called ok callback.
PASS Add Device callback called.
PASS navigator.getNetworkServices() called 2nd ok callback.
PASS srvs.servicesAvailable>=1 is true
PASS navigator.getNetworkServices() called eventCB callback.
PASS event.propertyset.length>0 is true
PASS event.friendlyName.length>0 is true
PASS event.serviceType.length>0 is true
PASS event.uuid.length>0 is true
PASS successfullyParsed is true

TEST COMPLETE
